// orthnova-offline/src/types.ts
export interface ServiceRow {
  id: string; // UUID
  service_name: string;
  service_type: string;
  price: number;
  created_at: string;
}